Fix weight loading for deprecate encoders
